refactor: Players가 Player 패키지에 속하도록 변경
