+++ src/main/java/domain/LadderGame.java
@@ -5,6 +5,7 @@ import domain.ladder.Height;
 import domain.ladder.Ladder;
 import domain.ladder.LadderFactory;
 import domain.ladder.Width;
+import domain.player.Players;
 import domain.player.Position;
 import java.util.ArrayList;
 import java.util.List;
