+++ src/main/java/domain/ladder/Width.java
@@ -1,6 +1,6 @@
 package domain.ladder;
 
-import domain.Players;
+import domain.player.Players;
 
 public record Width(int value) {
     public static Width from(final Players players) {
